Added jest test for Apollo server.
