+++ 3/package.json
@@ -4,13 +4,17 @@
   "description": "Test-drive the development of a GraphQL server.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "apollo-server": "^2.9.9",
+    "apollo-server-testing": "^2.9.9",
     "graphql": "^14.5.8"
+  },
+  "devDependencies": {
+    "jest": "^24.9.0"
   }
 }
